Implemented 2.3 and 2.4
Implemented methods for calculating net salary, calculating bonus salary, and getting the age of the employee. 

2.1 already seems completely implemented from Julia's end via DartAddEmployeeView, controller and view logic included.

2.2 seems like it is implemented as well from Julia's end but it does not have controller and view logic. 

Chose to not code the controller and view logic for 2.2 because I'm confused overall on the code given - needs a little clearing up.

In summary, everything is done besides rendering the view for Epic Feature 2.